chore: move unpivot from internal/debug to experimental
I needed to use unpivot in another repo, but since it's under
a directory called "internal" it isn't visible.
