+++ libflux/go/libflux/buildinfo.gen.go
@@ -154,8 +154,8 @@ var sourceHashes = map[string]string{
 	"stdlib/experimental/count_test.flux":                                                         "120a145eb287266e35638cc5220f5cecf0eee5bf2eec1bb1e9852e7d382050c7",
 	"stdlib/experimental/csv/csv.flux":                                                            "94ae72b5fb50b8d65e9b384aebaa890f04bc914bb344ea751a64f729b5c849fe",
 	"stdlib/experimental/distinct_test.flux":                                                      "6d2891556d985d16ac9bd33176dd527bb34246afa1ac81dde65f1335885b6b02",
-	"stdlib/experimental/experimental.flux":                                                       "426c8a101e86b9b7bb8b2d0b4410c0dc035711acf42ab02f67953b09ba8210bf",
-	"stdlib/experimental/experimental_test.flux":                                                  "206cbb54d35da80a6736dec990219f5bbabcdc3c65fbb9d1041058c8aae3450d",
+	"stdlib/experimental/experimental.flux":                                                       "1330a4983abad375a4079d2857a940f50364e510de98aeb808104b52d3727d06",
+	"stdlib/experimental/experimental_test.flux":                                                  "0fb9e187e7c90ccf9dffa41587dd6cfe255990142d11e1772b0a219d6f8a75b1",
 	"stdlib/experimental/fill_test.flux":                                                          "626568ae94d54c4d7639813191b814af104594a4a4d66fd1d7d76c1c48767fc3",
 	"stdlib/experimental/first_test.flux":                                                         "3bd1ff03bac6a45a3c525abb5ded3377f08195b6a5094caa42c8fb8b96aa6268",
 	"stdlib/experimental/geo/asTracks_test.flux":                                                  "3b33b115eacb39e04177d552cc0180f8969a75832bf457267baceb6d506a6321",
@@ -256,8 +256,8 @@ var sourceHashes = map[string]string{
 	"stdlib/influxdata/influxdb/to_test.flux":                                                     "fcf5253c42f7668988e6b951d4ec11c5ba754ee19ba1fcc67ec3e0b3c0d3d579",
 	"stdlib/influxdata/influxdb/v1/v1.flux":                                                       "8fdebd101bdd025b105fda3c7fac0c496372cb7027aa4df499c4ae5fa24bacad",
 	"stdlib/internal/boolean/boolean.flux":                                                        "1c0a1564911dd5f4c35c17efb32158ab31c51b20ac0f25175545e50aadc87c94",
-	"stdlib/internal/debug/debug.flux":                                                            "104485f6a8d6e1cd2fc8111367687fed4cef2a63b5a7657124571b89b0e2420c",
-	"stdlib/internal/debug/debug_test.flux":                                                       "063349cd314ed79906a96d84ed3c6829c734d1607e0bbae0654bdc9e22210e3c",
+	"stdlib/internal/debug/debug.flux":                                                            "0374cf0aee4d4fcf436bb682386b98cbd23242214ce4443f28907879afea6c39",
+	"stdlib/internal/debug/debug_test.flux":                                                       "673fa36e1fb36aa55ec2e76f08bf2600faa61c1904e4c0a7229b7680d1f0228f",
 	"stdlib/internal/gen/gen.flux":                                                                "493294b00b1c010b3665db3cb99cf47c82452a7215c7a12c91da03f072a9ecaa",
 	"stdlib/internal/gen/tables_test.flux":                                                        "14a3d3542aa0815bf3fb912b045c23161322acff731bfa04ce380c5da9866654",
 	"stdlib/internal/influxql/influxql.flux":                                                      "51739e7d9026cae356fdef79a4340d2bda7fc10d1daa1a7adeae06014f85049c",
